requirements.txt
telethon==1.34.0
gspread==5.12.4
google-auth==2.28.1
python-dotenv==1.0.1
schedule==1.2.1
pandas==2.2.1 
oauth2client
gspread-formatting